+++ lib/trainsh/detectors/target/env.rb
@@ -1,0 +1,11 @@
+require_relative '../target'
+
+module TrainSH
+  module Detectors
+    class EnvTarget < TargetDetector
+      def self.url
+        ENV['TARGET']
+      end
+    end
+  end
+end